delete second UserMapper
