+++ src/test/java/school/faang/user_service/mapper/user/UserMapperTest.java
@@ -1,17 +1,32 @@
 package school.faang.user_service.mapper.user;
 
 import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.mapstruct.factory.Mappers;
+import school.faang.user_service.dto.promotion.UserResponseDto;
 import school.faang.user_service.dto.user.UserDto;
 import school.faang.user_service.dto.user.UserResponseShortDto;
 import school.faang.user_service.entity.User;
+import school.faang.user_service.entity.promotion.PromotionTariff;
+import school.faang.user_service.entity.promotion.UserPromotion;
 
+import java.time.LocalDateTime;
 import java.util.List;
 
+import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static school.faang.user_service.util.promotion.PromotionFabric.getUser;
+import static school.faang.user_service.util.promotion.PromotionFabric.getUserPromotion;
 
 public class UserMapperTest {
+    private static final long USER_ID = 1;
+    private static final String USERNAME = "username";
+    private static final PromotionTariff TARIFF = PromotionTariff.STANDARD;
+    private static final LocalDateTime LOCAL_DATE_TIME = LocalDateTime.of(2000, 1, 1, 1,1);
+
+    private final UserMapper responseUserMapper = Mappers.getMapper(UserMapper.class);
+
     private UserMapper userMapper;
     private User user1;
     private User user2;
@@ -49,5 +64,15 @@ public class UserMapperTest {
         assertEquals(user1.getUsername(), userDtos.get(0).getUsername());
         assertEquals(user2.getUsername(), userDtos.get(1).getUsername());
 
+    }
+
+    @Test
+    @DisplayName("Test convert event to response event")
+    void testToDto() {
+        UserPromotion userPromotion = getUserPromotion(TARIFF, TARIFF.getNumberOfViews());
+        User user = getUser(USER_ID, USERNAME, List.of(userPromotion), LOCAL_DATE_TIME);
+        var responseDto = new UserResponseDto(USER_ID, USERNAME, TARIFF.toString(), TARIFF.getNumberOfViews(), LOCAL_DATE_TIME);
+
+        assertThat(responseUserMapper.toUserResponseDto(user)).isEqualTo(responseDto);
     }
 }
